% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnova.R
\name{getAnova}
\alias{getAnova}
\title{Perform ANOVA}
\usage{
getAnova(tse, formula, ...)
}
\arguments{
\item{tse}{A \code{SummarizedExperiment} object.}

\item{formula}{A formula where the LHS specifies the variable to test
(assay name, colData, or rowData variable) and the RHS specifies the
grouping variable, e.g., \code{counts ~ Group}.}

\item{...}{Additional arguments passed to \code{anova_test}.}
}
\value{
A \code{data.frame} with test results.
}
\description{
Performs ANOVA to compare values between two or more groups.
}
\examples{
data(peerj13075, package = "mia")
tse <- peerj13075
tse <- mia::agglomerateByRank(tse, "phylum")
res <- getAnova(tse, counts ~ Geographical_location)

}
