% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKruskal.R
\name{getKruskal}
\alias{getKruskal}
\title{Perform Kruskal-Wallis test}
\usage{
getKruskal(tse, formula, ...)
}
\arguments{
\item{tse}{A \code{SummarizedExperiment} object.}

\item{formula}{A formula where the LHS specifies the variable to test
(assay name, colData, or rowData variable) and the RHS specifies the
grouping variable, e.g., \code{counts ~ Group}.}

\item{...}{Additional arguments passed to \code{rstatix::kruskal_test}.}
}
\value{
A \code{data.frame} with test results.
}
\description{
Performs Kruskal-Wallis test to compare values between two or more groups.
}
\examples{
data(peerj13075, package = "mia")
tse <- peerj13075
tse <- mia::agglomerateByRank(tse, "phylum")
res <- getKruskal(tse, counts ~ Geographical_location)

}
