% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTTest.R
\name{getTTest}
\alias{getTTest}
\title{Perform t-test}
\usage{
getTTest(tse, formula, pair.by = NULL, ...)
}
\arguments{
\item{tse}{A \code{SummarizedExperiment} object.}

\item{formula}{A formula where the LHS specifies the variable to test
(assay name, colData, or rowData variable) and the RHS specifies the
grouping variable, e.g., \code{counts ~ Group}.}

\item{pair.by}{\code{Character scalar} or \code{NULL}. Column for pairing
samples in paired tests. (Default: \code{NULL})}

\item{...}{Additional arguments passed to \code{rstatix::t_test}.}
}
\value{
A \code{data.frame} with test results.
}
\description{
Performs t-test to compare values between two groups.
}
\examples{
data(peerj13075, package = "mia")
tse <- peerj13075
tse <- mia::agglomerateByRank(tse, "phylum")
res <- getTTest(tse, counts ~ Geographical_location)

}
