% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWilcoxonTest.R
\name{getWilcoxonTest}
\alias{getWilcoxonTest}
\title{Perform Wilcoxon rank-sum test}
\usage{
getWilcoxonTest(tse, formula, pair.by = NULL, ...)
}
\arguments{
\item{tse}{A \code{SummarizedExperiment} object.}

\item{formula}{A formula where the LHS specifies the variable to test
(assay name, colData, or rowData variable) and the RHS specifies the
grouping variable, e.g., \code{counts ~ Group}.}

\item{pair.by}{\code{Character scalar} or \code{NULL}. Column for pairing
samples in paired tests. (Default: \code{NULL})}

\item{...}{Additional arguments passed to \code{rstatix::wilcox_test}.}
}
\value{
A \code{data.frame} with test results.
}
\description{
Performs Wilcoxon rank-sum test (Mann-Whitney U) to compare values between
two groups.
}
\examples{
data(peerj13075, package = "mia")
tse <- peerj13075
tse <- mia::agglomerateByRank(tse, "phylum")
res <- getWilcoxonTest(tse, counts ~ Geographical_location)

}
